renaming package to components
