+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/components/MonthSelectionTopBar.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.common
+package br.com.rbrthmn.ui.financialcompanion.components
 
 import InfiniteHorizontalCircularList
 import InfiniteVerticalCircularList
