+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/incomedivisions/IncomeDivisionsScreen.kt
@@ -61,7 +61,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.common.MonthSelectionTopBar
+import br.com.rbrthmn.ui.financialcompanion.components.MonthSelectionTopBar
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
 import java.time.LocalDate
